backend/src/agents/gumwall-agent/requirements.txt
# GumWall AI Agent Dependencies

# Strands SDK - AI agent framework
strands-agents>=0.1.0

# Strands tools (required for MCP integration)
strands-agents-tools>=0.1.0

# Amazon Bedrock AgentCore - Runtime integration
bedrock-agentcore>=0.1.0

# AWS SDK for Python
boto3>=1.34.0
botocore[crt]>=1.34.0

# Model Context Protocol support
mcp>=1.0.0

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
